requirements.txt
python-docx>=1.1.0
pdfplumber>=0.11.0
pytesseract>=0.3.13
Pillow>=10.0.0
spacy>=3.8.0
